+++ RELEASENOTES.md
@@ -28,6 +28,8 @@ Release notes
         ([#10667](https://github.com/google/ExoPlayer/issues/10667)).
     *   Enforce minimum `compileSdkVersion` to avoid compilation errors
         ([#10684](https://github.com/google/ExoPlayer/issues/10684)).
+    *   Add `Player.getVideoSurfaceSize` that returns the size of the surface on
+        which the video is rendered.
 *   Downloads:
     *   Fix potential infinite loop in `ProgressiveDownloader` caused by
         simultaneous download and playback with the same `PriorityTaskManager`